requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pypdf==4.3.1
pdfplumber==0.11.4
reportlab==4.2.2
pdf2image==1.17.0
Pillow==10.4.0
PyJWT[cryptography]==2.8.0
httpx==0.27.0
pydantic-settings==2.3.4
